backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
openml
kaggle
sqlalchemy
celery[redis]
python-multipart
aiofiles
redis
python-dotenv 